aggregation4

+++ backend/src/controllers/order.ts
@@ -5,11 +5,19 @@ import NotFoundError from '../errors/not-found-error'
 import Order, { IOrder } from '../models/order'
 import Product, { IProduct } from '../models/product'
 import User from '../models/user'
-import {sanitizeAggregationFilters} from '../middlewares/sanitize/sanitizeAggregationParams'
+import { cleanHtml } from '../middlewares/sanitize/sanitizeHtml'
 import { sanitizeQueryParams } from '../middlewares/sanitize/sanitizeQueryParams'
+import { sanitizeSearch } from '../middlewares/sanitize/sanitizeSearch'
+import { sanitizeOrder } from '../middlewares/sanitize/sanitizeOrder'
+import { sanitizeAggregationFilters } from '../middlewares/sanitize/sanitizeAggregationParams'
 
 // eslint-disable-next-line max-len
 // GET /orders?page=2&limit=5&sort=totalAmount&order=desc&orderDateFrom=2024-07-01&orderDateTo=2024-08-01&status=delivering&totalAmountFrom=100&totalAmountTo=1000&search=%2B1
+
+enum Role {
+    Admin = 'admin',
+    User = 'customer'
+}
 
 export const getOrders = async (
     req: Request,
@@ -17,6 +25,87 @@ export const getOrders = async (
     next: NextFunction
 ) => {
     try {
+
+        const user=res.locals.user;
+        const safeQuery=sanitizeQueryParams(req.query);
+        const pageNumber = Math.max(1, parseInt(req.query.page as string) || 1);
+        const limitNumber = Math.min(10, Math.max(1, parseInt(req.query.limit as string) || 10));
+        const searchTerm = safeQuery.search;
+        const safeSearch = sanitizeSearch(searchTerm);
+        const searchRegex = new RegExp(safeSearch, 'i');
+        const searchNumber = Number(safeSearch);
+        const { limit, page, search, ...otherParams } = safeQuery;
+       
+
+        if (!user.roles.include(Role.Admin)){
+            //const safeQuery=sanitizeQueryParams(req.query);
+            
+
+            //const pageNumber = Math.max(1, parseInt(safeQuery.page as string) || 1);
+            //const limitNumber = Math.min(10, Math.max(1, parseInt(safeQuery.limit as string) || 10));
+
+            // Базовые фильтры для пользователя
+            const userFilters: FilterQuery<Partial<IOrder>> = { customer: user._id };
+            if (searchTerm && typeof searchTerm === 'string') {
+                
+                
+                
+                const products = await Product.find({ title: searchRegex });
+                const productIds: Types.ObjectId[] = products.map((product) => product._id as Types.ObjectId);
+                
+                const userOrders = await Order.find(userFilters).populate('products');
+                
+                let filteredOrders = userOrders.filter(order => {
+                    const matchesProduct = order.products.some(product => 
+                        productIds.some((id: Types.ObjectId) => id.equals(product._id))
+                    );
+                    const matchesOrderNumber = !Number.isNaN(searchNumber) && 
+                        order.orderNumber === searchNumber;
+                    return matchesProduct || matchesOrderNumber;
+                });
+                
+                // Пагинация
+                const totalOrders = filteredOrders.length;
+                const totalPages = Math.ceil(totalOrders / limitNumber);
+                filteredOrders = filteredOrders.slice(
+                    (pageNumber - 1) * limitNumber,
+                    pageNumber * limitNumber
+                );
+                
+                const sanitizedOrders = filteredOrders.map(sanitizeOrder);
+                
+                return res.status(200).json({
+                    orders: sanitizedOrders,
+                    pagination: {
+                        totalOrders,
+                        totalPages,
+                        currentPage: Number(pageNumber),
+                        pageSize: Number(limitNumber),
+                    },
+                });
+            }
+            
+            const userOrders = await Order.find(userFilters)
+                .populate(['customer', 'products'])
+                .skip((pageNumber - 1) * limitNumber)
+                .limit(limitNumber);
+            
+            const totalOrders = await Order.countDocuments(userFilters);
+            const totalPages = Math.ceil(totalOrders / limitNumber);
+            
+            const sanitizedOrders = userOrders.map(sanitizeOrder);
+            
+            return res.status(200).json({
+                orders: sanitizedOrders,
+                pagination: {
+                    totalOrders,
+                    totalPages,
+                    currentPage: Number(pageNumber),
+                    pageSize: Number(limitNumber),
+                },
+            });
+        }
+
         const {
             sortField = 'createdAt',
             sortOrder = 'desc',
@@ -25,14 +114,8 @@ export const getOrders = async (
             totalAmountTo,
             orderDateFrom,
             orderDateTo,
-            search,
-        } = req.query
+        } = safeQuery
 
-        const page=Math.max(1, parseInt(req.query.page as string)||1)
-        const limit=Math.min(10, parseInt(req.query.limit as string)||10)
-
-        const safeQuery=sanitizeQueryParams(req.query)
-        const searchTerm=safeQuery.search
         const filters: FilterQuery<Partial<IOrder>> = {}
 
         if (status) {
@@ -98,10 +181,6 @@ export const getOrders = async (
 
 
         if (searchTerm && typeof searchTerm === 'string') {
-            const safeSearch = searchTerm.replace(/[.*+?^${}()|[\]\\]/g, '\\$&');
-            const searchRegex = new RegExp(safeSearch, 'i')
-            const searchNumber = Number(safeSearch)
-
             const searchConditions: any[] = [{ 'products.title': searchRegex }]
 
             if (!Number.isNaN(searchNumber)) {
@@ -125,8 +204,8 @@ export const getOrders = async (
 
         aggregatePipeline.push(
             { $sort: sort },
-            { $skip: (Number(page) - 1) * Number(limit) },
-            { $limit: Number(limit) },
+            { $skip: (Number(pageNumber) - 1) * Number(limitNumber) },
+            { $limit: Number(limitNumber) },
             {
                 $group: {
                     _id: '$_id',
@@ -142,15 +221,17 @@ export const getOrders = async (
 
         const orders = await Order.aggregate(aggregatePipeline)
         const totalOrders = await Order.countDocuments(filters)
-        const totalPages = Math.ceil(totalOrders / Number(limit))
+        const totalPages = Math.ceil(totalOrders / Number(limitNumber))
+
+        const sanitizedOrders = orders.map(sanitizeOrder)
 
         res.status(200).json({
-            orders,
+            sanitizedOrders,
             pagination: {
                 totalOrders,
                 totalPages,
-                currentPage: Number(page),
-                pageSize: Number(limit),
+                currentPage: Number(pageNumber),
+                pageSize: Number(limitNumber),
             },
         })
     } catch (error) {
@@ -164,13 +245,19 @@ export const getOrdersCurrentUser = async (
     next: NextFunction
 ) => {
     try {
+
+        const safeQuery=sanitizeQueryParams(req.query);
         const userId = res.locals.user._id
-        const { search } = req.query
-        const page=Math.max(1, parseInt(req.query.page as string)||1)
-        const limit=Math.min(10, parseInt(req.query.limit as string)||10)
+        const pageNumber = Math.max(1, parseInt(req.query.page as string) || 1);
+        const limitNumber = Math.min(10, Math.max(1, parseInt(req.query.limit as string) || 10));
+        const searchTerm = safeQuery.search;
+        const safeSearch = sanitizeSearch(searchTerm);
+        const searchRegex = new RegExp(safeSearch, 'i');
+        const searchNumber = Number(safeSearch);
+        //const { limit, page, search, ...otherParams } = safeQuery;
         const options = {
-            skip: (Number(page) - 1) * Number(limit),
-            limit: Number(limit),
+            skip: (Number(pageNumber) - 1) * Number(limitNumber),
+            limit: Number(limitNumber),
         }
 
         const user = await User.findById(userId)
@@ -194,10 +281,10 @@ export const getOrdersCurrentUser = async (
 
         let orders = user.orders as unknown as IOrder[]
 
-        if (search) {
+        if (searchTerm && typeof searchTerm === 'string') {
             // если не экранировать то получаем Invalid regular expression: /+1/i: Nothing to repeat
-            const searchRegex = new RegExp(search as string, 'i')
-            const searchNumber = Number(search)
+            //const searchRegex = new RegExp(search as string, 'i')
+            //const searchNumber = Number(search)
             const products = await Product.find({ title: searchRegex })
             const productIds:Types.ObjectId[] = products.map((product) => product._id as Types.ObjectId)
 
@@ -216,17 +303,19 @@ export const getOrdersCurrentUser = async (
         }
 
         const totalOrders = orders.length
-        const totalPages = Math.ceil(totalOrders / Number(limit))
+        const totalPages = Math.ceil(totalOrders / Number(limitNumber))
 
         orders = orders.slice(options.skip, options.skip + options.limit)
 
+        const sanitizedOrders = orders.map(sanitizeOrder);
+
         return res.send({
-            orders,
+            sanitizedOrders,
             pagination: {
                 totalOrders,
                 totalPages,
-                currentPage: Number(page),
-                pageSize: Number(limit),
+                currentPage: Number(pageNumber),
+                pageSize: Number(limitNumber),
             },
         })
     } catch (error) {
@@ -241,6 +330,12 @@ export const getOrderByNumber = async (
     next: NextFunction
 ) => {
     try {
+
+        const orderNumber = Number(cleanHtml(req.params.orderNumber))
+        if (Number.isNaN(orderNumber)) {
+            return next(new BadRequestError('Невалидный номер заказа'))
+        }
+
         const order = await Order.findOne({
             orderNumber: req.params.orderNumber,
         })
@@ -251,7 +346,9 @@ export const getOrderByNumber = async (
                         'Заказ по заданному id отсутствует в базе'
                     )
             )
-        return res.status(200).json(order)
+        
+        const sanitizedOrder = sanitizeOrder(order);
+        return res.status(200).json(sanitizedOrder)
     } catch (error) {
         if (error instanceof MongooseError.CastError) {
             return next(new BadRequestError('Передан не валидный ID заказа'))
@@ -267,6 +364,11 @@ export const getOrderCurrentUserByNumber = async (
 ) => {
     const userId = res.locals.user._id
     try {
+        const orderNumber = Number(cleanHtml(req.params.orderNumber))
+        if (Number.isNaN(orderNumber)) {
+            return next(new BadRequestError('Невалидный номер заказа'))
+        }
+
         const order = await Order.findOne({
             orderNumber: req.params.orderNumber,
         })
@@ -283,7 +385,9 @@ export const getOrderCurrentUserByNumber = async (
                 new NotFoundError('Заказ по заданному id отсутствует в базе')
             )
         }
-        return res.status(200).json(order)
+
+        const sanitizedOrder = sanitizeOrder(order);
+        return res.status(200).json(sanitizeOrder)
     } catch (error) {
         if (error instanceof MongooseError.CastError) {
             return next(new BadRequestError('Передан не валидный ID заказа'))
@@ -302,11 +406,30 @@ export const createOrder = async (
         const basket: IProduct[] = []
         const products = await Product.find<IProduct>({})
         const userId = res.locals.user._id
+
+        const cleanOrderData: { [key: string]: any } ={}
+        for (let prop in req.body) {
+            cleanOrderData[prop]=cleanHtml(req.body[prop])
+        }
+        
         const { address, payment, phone, total, email, items, comment } =
-            req.body
+            cleanOrderData
+
+        const cleanPhone = sanitizeSearch(phone);
+        /* const { address, payment, phone, total, email, items, comment } =
+            req.body */
+
+        if (!Array.isArray(items) || items.length === 0) {
+            return next(new BadRequestError('Нет товаров для заказа'))
+        }
+
+        const validItems = items.filter((id: any) => Types.ObjectId.isValid(id))
+        if (validItems.length !== items.length) {
+            return next(new BadRequestError('Невалидные ID товаров'))
+        }
 
         items.forEach((id: Types.ObjectId) => {
-            const product = products.find((p) =>  (p._id as Types.ObjectId).equals(id))
+            const product = products.find((p) => (p._id as Types.ObjectId).equals(id))
             if (!product) {
                 throw new BadRequestError(`Товар с id ${id} не найден`)
             }
@@ -315,6 +438,7 @@ export const createOrder = async (
             }
             return basket.push(product)
         })
+
         const totalBasket = basket.reduce((a, c) => a + c.price, 0)
         if (totalBasket !== total) {
             return next(new BadRequestError('Неверная сумма заказа'))
@@ -324,7 +448,7 @@ export const createOrder = async (
             totalAmount: total,
             products: items,
             payment,
-            phone,
+            phone:cleanPhone,
             email,
             comment,
             customer: userId,
@@ -333,7 +457,8 @@ export const createOrder = async (
         const populateOrder = await newOrder.populate(['customer', 'products'])
         await populateOrder.save()
 
-        return res.status(200).json(populateOrder)
+        const sanitizedOrder=sanitizeOrder(populateOrder);
+        return res.status(200).json(sanitizedOrder)
     } catch (error) {
         if (error instanceof MongooseError.ValidationError) {
             return next(new BadRequestError(error.message))
@@ -349,9 +474,21 @@ export const updateOrder = async (
     next: NextFunction
 ) => {
     try {
-        const { status } = req.body
+        const orderNumber = Number(cleanHtml(req.params.orderNumber))
+        if (Number.isNaN(orderNumber)) {
+            return next(new BadRequestError('Невалидный номер заказа'))
+        }
+
+        const safeQuery=sanitizeQueryParams(req.body);
+        const { status } = safeQuery
+
+        const validStatuses = ['new', 'completed', 'cancelled', 'delivering']
+        if (!validStatuses.includes(status)) {
+            return next(new BadRequestError('Невалидный статус заказа'))
+        }
+
         const updatedOrder = await Order.findOneAndUpdate(
-            { orderNumber: req.params.orderNumber },
+            { orderNumber },
             { status },
             { new: true, runValidators: true }
         )
@@ -362,7 +499,9 @@ export const updateOrder = async (
                     )
             )
             .populate(['customer', 'products'])
-        return res.status(200).json(updatedOrder)
+        
+        const sanitizedOrder=sanitizeOrder(updateOrder)
+        return res.status(200).json(sanitizedOrder)
     } catch (error) {
         if (error instanceof MongooseError.ValidationError) {
             return next(new BadRequestError(error.message))
@@ -381,6 +520,13 @@ export const deleteOrder = async (
     next: NextFunction
 ) => {
     try {
+        const safeQuery=sanitizeQueryParams(req.params)
+        const { id } = safeQuery
+
+         if (!Types.ObjectId.isValid(id)) {
+            return next(new BadRequestError('Невалидный ID заказа'))
+        }
+
         const deletedOrder = await Order.findByIdAndDelete(req.params.id)
             .orFail(
                 () =>
@@ -389,7 +535,9 @@ export const deleteOrder = async (
                     )
             )
             .populate(['customer', 'products'])
-        return res.status(200).json(deletedOrder)
+
+        const sanitizedOrder=sanitizeOrder(deleteOrder);    
+        return res.status(200).json(sanitizedOrder)
     } catch (error) {
         if (error instanceof MongooseError.CastError) {
             return next(new BadRequestError('Передан не валидный ID заказа'))
